error clear

+++ src/main/java/com/hireon/backend/Repository/DriveRepo.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface DriveRepo extends JpaRepository<Drive,Integer> {
+public interface DriveRepo extends JpaRepository<Drive,Long> {
 
 }
